+++ packages/action-menu/src/ActionMenu.ts
@@ -30,6 +30,7 @@ import actionMenuStyles from './action-menu.css.js';
  * @slot - menu items to be listed in the Action Menu
  * @slot icon - The icon to use for Action Menu
  * @slot label - The label to use on for the Action Menu
+ * @slot tooltip - Tooltip to use on for the Action menu
  * @attr selects - By default `sp-action-menu` does not manage a selection. If
  *   you'd like for a selection to be held by the `sp-menu` that it presents in
  *   its overlay, use `selects="single" to activate this functionality.
@@ -55,6 +56,7 @@ export class ActionMenu extends ObserveSlotText(PickerBase, 'label') {
                     <sp-icon-more class="icon"></sp-icon-more>
                 </slot>
                 <slot name="label" ?hidden=${!this.hasLabel}></slot>
+                <slot name="tooltip"></slot>
             `,
         ];
     }