+++ packages/swatch/src/SwatchGroup.ts
@@ -32,6 +32,7 @@ import type {
 } from './Swatch.js';
 
 export type SwatchGroupSizes = Exclude<ElementSize, 'xxs' | 'xl' | 'xxl'>;
+export type SwatchSelects = 'single' | 'multiple' | undefined;
 
 /**
  * @element sp-swatch-group
@@ -52,21 +53,10 @@ export class SwatchGroup extends SizedMixin(SpectrumElement, {
     public rounding: SwatchRounding;
 
     @property({ type: Array })
-    public get selected(): string[] {
-        return this._selected;
-    }
-
-    public set selected(selected: string[]) {
-        if (selected === this.selected) return;
-        const oldSelected = this.selected;
-        this._selected = selected;
-        this.requestUpdate('selected', oldSelected);
-    }
-
-    private _selected: string[] = [];
+    public selected: string[] = [];
 
     @property()
-    public selects: 'single' | 'multiple' | undefined;
+    public selects: SwatchSelects;
 
     private selectedSet = new Set<string>();
 
@@ -140,7 +130,7 @@ export class SwatchGroup extends SizedMixin(SpectrumElement, {
                 this.selectedSet.delete(target.value);
             }
         }
-        this._selected = [...this.selectedSet];
+        this.selected = [...this.selectedSet];
         const applyDefault = this.dispatchEvent(
             new Event('change', {
                 cancelable: true,
@@ -168,7 +158,7 @@ export class SwatchGroup extends SizedMixin(SpectrumElement, {
                 this.selectedSet.delete(value);
             }
         });
-        this._selected = [...this.selectedSet];
+        this.selected = [...this.selectedSet];
     };
 
     private getPassthroughSwatchActions(
@@ -179,7 +169,14 @@ export class SwatchGroup extends SizedMixin(SpectrumElement, {
             rounding?: SwatchRounding;
             shape?: SwatchShape;
             size?: SwatchGroupSizes;
+            selects?: SwatchSelects;
         } = {};
+        if (
+            changes.has('selects') &&
+            (this.selects || typeof changes.get('selects') !== 'undefined')
+        ) {
+            targetValues.selects = this.selects;
+        }
         if (
             changes.has('border') &&
             (this.border || typeof changes.get('border') !== 'undefined')
@@ -207,6 +204,22 @@ export class SwatchGroup extends SizedMixin(SpectrumElement, {
         const passThroughSwatchActions: ((swatch: Swatch) => void)[] = [];
         if (Object.keys(targetValues).length) {
             passThroughSwatchActions.push((swatch) => {
+                if (window.__swc.DEBUG) {
+                    if (
+                        'selects' in targetValues &&
+                        targetValues.selects !== 'multiple' &&
+                        swatch.mixedValue
+                    ) {
+                        window.__swc.warn(
+                            this,
+                            `<sp-swatch> elements can only leverage the "mixed-value" attribute when their <sp-swatch-group> parent element is also leveraging "selects="multiple""`,
+                            'https://opensource.adobe.com/spectrum-web-components/components/swatch-group/#multiple',
+                            {
+                                type: 'accessibility',
+                            }
+                        );
+                    }
+                }
                 if ('border' in targetValues)
                     swatch.border = targetValues.border;
                 if ('rounding' in targetValues)
@@ -264,7 +277,10 @@ export class SwatchGroup extends SizedMixin(SpectrumElement, {
         if (changes.has('selected')) {
             swatchActions.push((swatch) => {
                 currentValues.add(swatch.value);
-                if (nextSelected.has(swatch.value) || swatch.selected) {
+                if (
+                    nextSelected.has(swatch.value) ||
+                    (!this.hasUpdated && swatch.selected)
+                ) {
                     swatch.selected = true;
                 } else {
                     swatch.selected = false;